pyproject.toml
[project]
name = "pokequiz"
version = "0.1.0"
description = "Multi-mode Pokemon quiz game powered by PokeAPI/Pokebase"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pokebase>=1.4.1",
  "pillow>=10.0.0",
]

[project.scripts]
pokequiz = "pokequiz.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
